backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
requests==2.31.0
